backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
daphne==4.0.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
